Traffic violation detection/AI 모델 소스코드/nia-82-134-main/requirements.txt
# with conda
# cudatoolkit=11.0, cudnn=8
# with pip
-f https://download.pytorch.org/whl/torch_stable.html
-f https://download.openmmlab.com/mmcv/dist/cu110/torch1.7.0/index.html
colorlog==6.6.0
easydict==1.9
future==0.18.2
ipykernel
ipywidgets==7.6.5
jupyter==1.0.0
matplotlib==3.5.0
mmcv-full
mmdet==2.18.1
omegaconf==2.1.1
pandas==1.3.4
scikit-learn==1.0.1
tensorboard==2.7.0
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
tqdm==4.62.3
